requirements.txt
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
